+++ frontend/src/components/UI/ProfileComponents/Buttons/MessageButton.tsx
@@ -1,0 +1,7 @@
+import Button from "./Button";
+
+function MessageButton() {
+  return <Button text="message" onClick={() => alert("message")} />;
+}
+
+export default MessageButton;
