+++ frontend/src/components/UI/ProfileComponents/Buttons/MoreOptions.tsx
@@ -1,0 +1,6 @@
+import Button from "./Button";
+
+function MoreOptionsButton() {
+  return <Button text="..." />;
+}
+export default MoreOptionsButton;
