+++ frontend/src/pages/ProfilePage.tsx
@@ -1,50 +1,12 @@
 import type { ReactNode } from "react";
 import { useState } from "react";
 import { NavLink, Route, Routes } from "react-router";
-import ContentContainer from "../components/UI/ProfileComponents/ContentContainer";
-import Tabs from "../components/UI/ProfileComponents/Tabs";
-import profileData from "../templates/ProfileTemplate";
-
-interface ButtonProps {
-  text: string;
-  onClick?: (event: React.MouseEvent<HTMLButtonElement>) => void;
-  className?: string;
-}
-
-function Button(props: ButtonProps) {
-  return (
-    <button
-      onClick={props.onClick}
-      className={`${props.className} py-1 font-bold px-2 items-center justify-center border border-accent-secondary rounded-md`}
-      type="button"
-    >
-      <span>{props.text}</span>
-    </button>
-  );
-}
-
-interface FollowButtonProps {
-  onFollow: (event: React.MouseEvent<HTMLButtonElement>) => void;
-  followStatus: boolean;
-}
-
-function FollowButton(props: FollowButtonProps) {
-  return (
-    <Button
-      onClick={props.onFollow}
-      text={props.followStatus ? "followed" : "follow"}
-      className="w-24"
-    />
-  );
-}
-
-function MessageButton() {
-  return <Button text="message" onClick={() => alert("message")} />;
-}
-
-function MoreOptions() {
-  return <Button text="..." />;
-}
+import FollowButton from "../components/UI/Buttons/FollowButton.tsx";
+import MessageButton from "../components/UI/ProfileComponents/Buttons/MessageButton.tsx";
+import MoreOptionsButton from "../components/UI/ProfileComponents/Buttons/MoreOptions.tsx";
+import ContentContainer from "../components/UI/ProfileComponents/ContentContainer.tsx";
+import Tabs from "../components/UI/ProfileComponents/Tabs.tsx";
+import profileData from "../templates/ProfileTemplate.tsx";
 
 function ProfileActions() {
   const [isFollowed, setIsFollowed] = useState(false);
@@ -56,7 +18,7 @@ function ProfileActions() {
     <div className="flex gap-3 items-center">
       <FollowButton onFollow={handleOnFollow} followStatus={isFollowed} />
       <MessageButton />
-      <MoreOptions />
+      <MoreOptionsButton />
     </div>
   );
 }
@@ -168,17 +130,17 @@ export default function ProfilePage() {
               element={
                 <ContentContainer>
                   {profileData.posts.map((post) => (
-                    <div key={post.id} className="group h-64">
-                      <a
-                        href={`/${profileData.userData.userName}/posts/${post.id}`}
-                      >
-                        <img
-                          className="rounded-sm group-hover:opacity-40 transition duration-150 ease-in-out"
-                          src={post.postMedia}
-                          alt="post media"
-                        />
-                      </a>
-                    </div>
+                    <a
+                      key={post.id}
+                      className="group h-96 flex"
+                      href={`/${profileData.userData.userName}/posts/${post.id}`}
+                    >
+                      <img
+                        className="rounded-sm w-full object-cover group-hover:opacity-40 transition duration-150 ease-in-out"
+                        src={post.postMedia}
+                        alt="post media"
+                      />
+                    </a>
                   ))}
                 </ContentContainer>
               }
